requirements.txt
numpy
biopython
scikit-learn
progressbar2